Refactor trace store legacy limit helpers
